RIP-373, break up canvas_utils (eg, egrades to dedicated utils) to stay <700 lines
