+++ ripley/api/canvas_site_controller.py
@@ -89,7 +89,7 @@ def create_project_site():
         if not name or len(name) > 255:
             raise BadRequestError(f"'Invalid project site name: '{name}'")
         project_site = create_canvas_project_site(name=name, owner_uid=current_user.uid)
-        return tolerant_jsonify(project_site)
+        return tolerant_jsonify(canvas_site_to_api_json(project_site))
     else:
         app.logger.warning(f'Unauthorized request to {request.path}')
         return app.login_manager.unauthorized()
